crewai/trip_planner/pyproject.toml
[project]
name = "trip-planner-crew"
version = "0.1.0"
description = "Simple AI trip planner"
authors = [{name = "Joao Moura, Tevin Wright", email = "tevinswright@gmail.com"}]
requires-python = ">=3.10.0,<3.12"
dependencies = [
    "crewai>=0.152.0",
    "unstructured==0.10.25",
    "pyowm==3.3.0",
    "tools>=0.1.9",
    "python-dotenv==1.0.0",
    "langchain (>=1.1.0,<2.0.0)",
    "langchain-openai (>=1.1.0,<2.0.0)",
]

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"